Title: 숫자 사이의 배수, Time: 139ms, Memory: 10MB, Status: Passed - CodeTree

+++ 231217/숫자 사이의 배수/multiplication-between-numbers.java
@@ -16,7 +16,7 @@ public class Main {
             }
 
         }
-        float k = sum/cnt;
+        float k = (float)sum/cnt;
 
         System.out.print(sum+" ");
         System.out.printf("%.1f", k);
